backend/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
PyYAML==6.0.2
oracledb==2.2.1
SQLAlchemy>=2.0
alembic>=1.13
bcrypt>=4.1
oci==2.131.1
pydantic-settings
langchain-core>=0.3.10,<0.4.0
langchain-community>=0.3.10,<0.4.0
spacy>=3.7.0
detect-secrets>=1.4.0
pdfplumber>=0.10.3
tqdm>=4.66.0
